+++ packages/orq-rc/src/orq_ai_sdk/models/deployments.py
@@ -187,7 +187,7 @@ r"""The contents of the assistant message. Required unless `tool_calls` or `func
 
 
 DeploymentsPrefixMessages4Role = Literal["assistant",]
-r"""The role of the messages author, in this case `assistant` or `exception`."""
+r"""The role of the messages author, in this case `assistant`."""
 
 
 class AudioTypedDict(TypedDict):
@@ -243,7 +243,7 @@ class ToolCalls(BaseModel):
 
 class AssistantMessageTypedDict(TypedDict):
     role: DeploymentsPrefixMessages4Role
-    r"""The role of the messages author, in this case `assistant` or `exception`."""
+    r"""The role of the messages author, in this case `assistant`."""
     content: NotRequired[Nullable[PrefixMessagesContentTypedDict]]
     r"""The contents of the assistant message. Required unless `tool_calls` or `function_call` is specified."""
     refusal: NotRequired[Nullable[str]]
@@ -264,7 +264,7 @@ class AssistantMessageTypedDict(TypedDict):
 
 class AssistantMessage(BaseModel):
     role: DeploymentsPrefixMessages4Role
-    r"""The role of the messages author, in this case `assistant` or `exception`."""
+    r"""The role of the messages author, in this case `assistant`."""
 
     content: OptionalNullable[PrefixMessagesContent] = UNSET
     r"""The contents of the assistant message. Required unless `tool_calls` or `function_call` is specified."""
@@ -736,7 +736,7 @@ r"""The contents of the assistant message. Required unless `tool_calls` or `func
 
 
 DeploymentsMessages4Role = Literal["assistant",]
-r"""The role of the messages author, in this case `assistant` or `exception`."""
+r"""The role of the messages author, in this case `assistant`."""
 
 
 class MessagesAudioTypedDict(TypedDict):
@@ -792,7 +792,7 @@ class MessagesToolCalls(BaseModel):
 
 class MessagesAssistantMessageTypedDict(TypedDict):
     role: DeploymentsMessages4Role
-    r"""The role of the messages author, in this case `assistant` or `exception`."""
+    r"""The role of the messages author, in this case `assistant`."""
     content: NotRequired[Nullable[DeploymentsMessagesContentTypedDict]]
     r"""The contents of the assistant message. Required unless `tool_calls` or `function_call` is specified."""
     refusal: NotRequired[Nullable[str]]
@@ -813,7 +813,7 @@ class MessagesAssistantMessageTypedDict(TypedDict):
 
 class MessagesAssistantMessage(BaseModel):
     role: DeploymentsMessages4Role
-    r"""The role of the messages author, in this case `assistant` or `exception`."""
+    r"""The role of the messages author, in this case `assistant`."""
 
     content: OptionalNullable[DeploymentsMessagesContent] = UNSET
     r"""The contents of the assistant message. Required unless `tool_calls` or `function_call` is specified."""